Separate TESS logic into child class
